Improve a hacky and incorrect query
The query that finds the matching numbers was incorrect.
When the bit numbers representation ends with `0`(having 49)
the DB query spliting the bits into array of ones(`E'0*'`) will
end up with array of ones and ending with an empty string -> {1,1,.., ''}.

The new query that remove the zeros after the bitwise AND is simpler and
works correctly for the above example.